refactor: add DashboardLayout with shared sidebar
