+++ src/App.tsx
@@ -1,15 +1,18 @@
-import { DashboardUser } from "./pages/DashboardUser";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
-import { Skills } from "./pages/Skills";
-import { Profil } from "./pages/Profil";
+import { DashboardUser } from "@/pages/dashboard/DashboardUser";
+import { Skills } from "@/pages/dashboard/Skills";
+import { Profil } from "@/pages/dashboard/Profil";
+import { DashboardLayout } from "@/pages/dashboard/DashboardLayout";
 
 function App() {
   return (
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={<DashboardUser />} />
-        <Route path="/skills" element={<Skills />} />
-        <Route path="/profil" element={<Profil />} />
+        <Route path="/" element={<DashboardLayout />}>
+          <Route index element={<DashboardUser />} />
+          <Route path="skills" element={<Skills />} />
+          <Route path="profil" element={<Profil />} />
+        </Route>
       </Routes>
     </BrowserRouter>
   );
