--- src/pages/Profil.tsx
@@ -1,12 +1,0 @@
-import { Sidebar } from "@/component/sidebar/Sidebar";
-
-export const Profil = () => {
-  return (
-    <div>
-      <Sidebar />
-      <div className="flex items-center justify-center h-screen">
-        <h1 className="text-3xl font-bold">Profil page</h1>
-      </div>
-    </div>
-  );
-};
