--- src/pages/Skills.tsx
@@ -1,12 +1,0 @@
-import { Sidebar } from "@/component/sidebar/Sidebar";
-
-export const Skills = () => {
-  return (
-    <div>
-      <Sidebar />
-      <div className="flex items-center justify-center h-screen">
-        <h1 className="text-3xl font-bold">Skills listing page</h1>
-      </div>
-    </div>
-  );
-};
