+++ src/pages/dashboard/DashboardLayout.tsx
@@ -1,0 +1,11 @@
+import { Sidebar } from "@/component/sidebar/Sidebar";
+import { Outlet } from "react-router-dom";
+
+export const DashboardLayout = () => {
+  return (
+    <div>
+      <Sidebar />
+      <Outlet />
+    </div>
+  );
+};
