+++ src/pages/dashboard/DashboardUser.tsx
@@ -1,0 +1,7 @@
+export const DashboardUser = () => {
+  return (
+    <div className="flex items-center justify-center h-screen">
+      <h1 className="text-3xl font-bold">Dashboard user</h1>
+    </div>
+  );
+};
