+++ src/pages/dashboard/Profil.tsx
@@ -1,0 +1,7 @@
+export const Profil = () => {
+  return (
+    <div className="flex items-center justify-center h-screen">
+      <h1 className="text-3xl font-bold">Profil page</h1>
+    </div>
+  );
+};
