+++ src/pages/dashboard/Skills.tsx
@@ -1,0 +1,7 @@
+export const Skills = () => {
+  return (
+    <div className="flex items-center justify-center h-screen">
+      <h1 className="text-3xl font-bold">Skills listing page</h1>
+    </div>
+  );
+};
