--- src/pages/home.tsx
@@ -1,8 +1,0 @@
-export const Home = () => {
-  return (
-    <div>
-      <h1>Welcome to the Home Page</h1>
-      <p>This is the main page of the application.</p>
-    </div>
-  );
-};
